+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 Future minor version (release soon)
 -----------------------------------
 
-* Fix bug in progress bar for small number of samples.
+* Remove import of matplotlib backend QtAgg and related error message, except for sarvey_plot.
 
 1.3.0 (2026-02-23)
 ------------------